Added unit test function: void test_if_end_game(void)
